refactor post meta example block
