+++ example/src/blocks/post-meta/block.json
@@ -8,5 +8,6 @@
 			"type": "string",
 			"default": ""
 		}
-	}
+	},
+	"editorScript": "file:./index.js"
 }
